requirements.txt
PyRSMQ==0.6.0
redis==5.0.7
pydantic==2.11.0
